refactor: move from internal_key_handler to new
