+++ test/new/service/shortcuts/command_shortcut_events/arrow_keys_handler_test.dart
@@ -4,7 +4,7 @@ import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../new/infra/testable_editor.dart';
+import '../../../infra/testable_editor.dart';
 
 void main() async {
   TestWidgetsFlutterBinding.ensureInitialized();
