+++ test/new/service/shortcuts/command_shortcut_events/copy_paste_handler_test.dart
@@ -4,8 +4,8 @@ import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../infra/clipboard_test.dart';
-import '../../new/infra/testable_editor.dart';
+import '../../../../infra/clipboard_test.dart';
+import '../../../infra/testable_editor.dart';
 
 void main() async {
   late MockClipboard mockClipboard;
@@ -36,12 +36,12 @@ void main() async {
         (tester) async {
       await _testSameNodeCopyPaste(tester, Document.fromJson(paragraphData));
     });
-    testWidgets('Presses Command + A in nested docment and copy text',
+    testWidgets('Presses Command + A in nested document and copy text',
         (tester) async {
       await _testHandleCopy(tester, Document.fromJson(data));
     });
     testWidgets(
-        'Presses Command + A in nested docment and copy text nestednode',
+        'Presses Command + A in nested document and copy text nestednode',
         (tester) async {
       // TODO: fix this test
       // await _testNestedNodeCopyPaste(tester, Document.fromJson(exampledoc));
