+++ test/new/service/shortcuts/command_shortcut_events/exit_editing_mode_handler_test.dart
@@ -1,7 +1,7 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import '../../new/infra/testable_editor.dart';
+import '../../../infra/testable_editor.dart';
 
 void main() async {
   setUpAll(() {
@@ -33,8 +33,6 @@ void main() async {
         ),
       );
     });
-
-    // Future<void> _testSelection()
   });
 }
