+++ test/new/service/shortcuts/command_shortcut_events/redo_undo_handler_test.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import '../../new/infra/testable_editor.dart';
-import '../../new/util/util.dart';
+import '../../../infra/testable_editor.dart';
+import '../../../util/util.dart';
 
 void main() async {
   setUpAll(() {
